# apps/backend/src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "2.0.0",
    "title": "Lumisovellus API",
    "description": "Unified backend API for Lumisovellus - Snow condition tracking and rescue coordination system",
    "contact": {
      "name": "Mindhive Oy",
      "email": "support@mindhive.fi"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token (without the Bearer prefix)"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/segments": {
      "get": {
        "tags": ["Segments"],
        "summary": "Get all segments",
        "description": "Retrieve all ski segments with their coordinates and terrain information. Supports filtering by bounding box, search, and updatedSince.",
        "parameters": [
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "Bounding box to filter segments (format: minLat,minLng,maxLat,maxLng)",
            "example": "64.0,25.0,66.0,30.0"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "Search term to filter segments by name",
            "example": "tunturi"
          },
          {
            "name": "updatedSince",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time",
            "description": "Return only segments updated since this date (ISO 8601 format)",
            "example": "2024-01-01T00:00:00Z"
          }
        ],
        "responses": {
          "200": {
            "description": "Segments retrieved successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Segment"
                  }
                },
                "meta": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/api/v1/segments/{id}/updates": {
      "get": {
        "tags": ["Segments"],
        "summary": "Get latest updates for a segment",
        "description": "Retrieve the most recent updates for a specific segment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid",
            "description": "Segment ID (UUID)"
          }
        ],
        "responses": {
          "200": {
            "description": "Updates retrieved successfully",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/api/v1/updates": {
      "get": {
        "tags": ["Updates"],
        "summary": "Get updates for segments",
        "description": "Get updates filtered by updatedSince or time range; include review details.",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 30,
              "default": 3
            },
            "description": "Number of days to look back (ignored if updatedSince/from/to provided)"
          },
          {
            "name": "segmentId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter updates by a specific segment ID"
          },
          {
            "name": "updatedSince",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Return updates since this timestamp (ISO 8601)"
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Start of time range (ISO 8601). If provided, overrides days/updatedSince."
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "End of time range (ISO 8601). If provided, overrides days/updatedSince."
          }
        ],
        "responses": {
          "200": {
            "description": "Updates retrieved successfully, with review details included",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "timestamp": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  }
}